function implementaion

+++ src/lib/math_lib.h
@@ -6,7 +6,7 @@ double Sub(double x, double y);
 double Div(double x, double y);
 double Mul(double x, double y);
 unsigned long long Factorial(unsigned int x);
-double Root(double x, double y);
+double Root(double x, unsigned int y);
 double Pow(double x, unsigned int y);
 double Abs(double x);
